m4-document-generation/requirements.txt
Flask==2.3.2
gunicorn==20.1.0
python-docx==1.1.2
requests==2.32.3
python-dotenv==1.0.0
